PR-1: tabbed-app shell — extract Today screen, add 6-tab bar
Reshapes the single-screen home into a tab shell per the
tabbed-app spec. PR-1 is the structural change only — all
existing functionality moves under the Today tab; the other
five tabs (Body, Move, Mind, Places, Roles) render placeholder
"Coming soon" screens that subsequent PRs fill in.

App.tsx shrinks from ~2500 to ~1980 lines. ALL data effects
(HealthKit queries, location tracking, CloudKit sync, deep-link
routing, OTA checks, the module-scope TaskManager.defineTask
registration) stay in App.tsx unchanged — only the render JSX
is extracted into screens/TodayScreen.tsx. App.tsx now wires
data + handlers to whichever screen the active tab selects,
and renders the eight existing modals (About, Settings,
Affirmation, Grateful, Journal, LocationDetail, MetricDetail,
WorkoutAnalysis) at app-level above the TabBar.

Spec: docs/superpowers/specs/2026-05-25-tabbed-app-design.md
Plan: docs/superpowers/plans/2026-05-25-tabbed-app-plan.md

New files
- components/TabBar.tsx — hand-rolled 6-tab bottom bar; testID per tab
- components/MetricCard.tsx — extracted from inline App.tsx so TodayScreen can import it
- lib/appTypes.ts — ContextSnapshot + LocationData types (avoids circular import App↔Today)
- screens/TodayScreen.tsx — header + reflect + metric grid + share footer
- screens/{Body,Move,Mind,Places,Roles}Screen.tsx — "Coming soon" stubs

Tests
- 4 new tab-navigation tests (renders, all six tabs, default Today, switch to Body)
- All 446 tests pass (23 suites). Typecheck clean.

GymTimer early-exit is preserved as-is (line ~1717) — its
migration to the Move tab is PR-3 scope. LocationDetailSheet
moved out of the ScrollView to app-level alongside the other
modals (its prop API already accepts null location).
